+++ src/strojure/undertow/handler/csp.clj
@@ -19,34 +19,30 @@
 ;;,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
 
 (defn report-uri-handler
-  "Handles CSP report URI and invokes `:report-callback` function with
-  `HttpServerExchange` as argument. Respond with HTTP 200. Used by
-  [[csp-handler]] when `:report-callback` option is defined.
+  "Handles CSP report URI with HttpHandler `:report-handler`. Used by
+  [[csp-handler]] when `:report-handler` option is defined.
 
   Configuration map keys:
 
-  - `:report-callback` – a function `(fn callback [exchange] ...)`.
-      + Required.
-      + Invoked when request URI equals `:report-uri`.
-      + The return value is ignored.
+  - `:report-handler` – the HttpHandler to be applied when request URI equals
+    `:report-uri`, required.
 
   - `:report-uri` – a string with request `:uri` to match for.
       + Exact value is matched.
       + Default value is \"/csp-report\".
   "
   {:tag HttpHandler :added "1.1"}
-  [next-handler {:keys [report-callback, report-uri]}]
-  (assert (fn? report-callback))
-  (let [report-uri (or report-uri report-uri-default)
-        next-handler (types/as-handler next-handler)]
+  [next-handler {:keys [report-uri, report-handler]}]
+  (let [next-handler (types/as-handler next-handler)
+        report-handler (types/as-handler report-handler)
+        report-uri (or report-uri report-uri-default)]
     (assert (string? report-uri) (str "Expect string in `:report-uri`: " report-uri))
     (reify HttpHandler
       (handleRequest [_ e]
-        (if (.equals ^String report-uri (.getRequestURI e))
-          (doto e (report-callback)
-                  (.setStatusCode 200)
-                  (.endExchange))
-          (.handleRequest next-handler e))))))
+        (-> (if (.equals ^String report-uri (.getRequestURI e))
+              report-handler
+              next-handler)
+            (.handleRequest e))))))
 
 ;;,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
 
@@ -72,18 +68,17 @@
   - `:random-nonce-fn` – optional 0-arity function to generate nonce for every
                          request.
 
-  - `:report-callback` – a function `(fn callback [exchange] ...)` to handle
-    `report-uri` directive.
-      + When presented then handler is wrapped with [[report-uri-handler]].
+  - `::report-handler` – a HttpHandler to handle `report-uri` directive.
+      + When presented then `next-handler` is wrapped with [[report-uri-handler]].
       + If policy map does not have `report-uri` directive then it is added with
         default value \"/csp-report\".
   "
   {:tag HttpHandler :added "1.1"}
-  [next-handler {:keys [policy, report-only, random-nonce-fn, report-callback]}]
+  [next-handler {:keys [policy, report-only, random-nonce-fn, report-handler]}]
   (let [next-handler (types/as-handler next-handler)
         header-name (HttpString. ^String (csp/header-name report-only))
-        report-uri (and report-callback (csp/find-directive :report-uri policy))
-        policy (cond-> policy (and report-callback (not report-uri))
+        report-uri (and report-handler (csp/find-directive :report-uri policy))
+        policy (cond-> policy (and report-handler (not report-uri))
                               (assoc :report-uri report-uri-default))
         header-value-fn (csp/header-value-fn policy)
         nonce-fn (and (csp/requires-nonce? header-value-fn)
@@ -102,8 +97,8 @@
                        (doto (.getResponseHeaders e)
                          (.put header-name ^String (header-value-fn)))
                        (-> next-handler (.handleRequest e)))))
-      report-callback
-      (report-uri-handler {:report-callback report-callback
+      report-handler
+      (report-uri-handler {:report-handler report-handler
                            :report-uri report-uri}))))
 
 (defn get-request-nonce