+++ test/strojure/undertow/handler/csp_test.clj
@@ -1,11 +1,12 @@
 (ns strojure.undertow.handler.csp-test
   (:require [clojure.test :as test :refer [deftest testing]]
             [java-http-clj.core :as http]
+            [strojure.undertow.api.exchange :as exchange]
             [strojure.undertow.api.types :as types]
             [strojure.undertow.handler :as handler]
             [strojure.undertow.handler.csp :as csp]
             [strojure.undertow.server :as server])
-  (:import (io.undertow.server HttpHandler HttpServerExchange)))
+  (:import (io.undertow.server HttpHandler)))
 
 (set! *warn-on-reflection* true)
 
@@ -29,52 +30,76 @@
 
 ;;,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
 
-(def ^{:arglists '([{:keys [report-callback, report-uri]}])
+(def ^{:arglists '([{:keys [report-handler, report-uri]}])
        :private true}
   report-uri-handler* (partial csp/report-uri-handler dummy-handler))
 
+(defn- with-request-body-string
+  [f]
+  (handler/dispatch
+    (reify HttpHandler
+      (handleRequest [_ e]
+        (-> (exchange/get-input-stream e)
+            (.readAllBytes)
+            (String.)
+            (f))))))
+
 (deftest report-uri-handler-t
 
-  (test/is (= "/csp-report"
-              (let [a! (atom :undefined)
-                    callback (fn [^HttpServerExchange e] (reset! a! (.getRequestURI e)))]
-                (exec {:handler (report-uri-handler* {:report-callback callback})
-                       :uri "/csp-report"})
-                @a!)))
+  (testing "URI matching"
 
-  (test/is (= "/custom-csp-report"
-              (let [a! (atom :undefined)
-                    callback (fn [^HttpServerExchange e] (reset! a! (.getRequestURI e)))]
-                (exec {:handler (report-uri-handler* {:report-uri "/custom-csp-report"
-                                                      :report-callback callback})
-                       :uri "/custom-csp-report"})
-                @a!)))
+    (test/is (= "/csp-report"
+                (let [a! (atom :undefined)
+                      handler (reify HttpHandler (handleRequest [_ e] (reset! a! (.getRequestURI e))))]
+                  (exec {:handler (report-uri-handler* {:report-handler handler})
+                         :uri "/csp-report"})
+                  @a!)))
 
-  (test/is (= :undefined
-              (let [a! (atom :undefined)
-                    callback (fn [^HttpServerExchange e] (reset! a! (.getRequestURI e)))]
-                (exec {:handler (report-uri-handler* {:report-callback callback})
-                       :uri "/not-csp-report"})
-                @a!)))
+    (test/is (= "/custom-csp-report"
+                (let [a! (atom :undefined)
+                      handler (reify HttpHandler (handleRequest [_ e] (reset! a! (.getRequestURI e))))]
+                  (exec {:handler (report-uri-handler* {:report-uri "/custom-csp-report"
+                                                        :report-handler handler})
+                         :uri "/custom-csp-report"})
+                  @a!)))
 
-  (test/is (= #_:status 200
-                        (let [a! (atom :undefined)
-                              callback (fn [^HttpServerExchange e] (reset! a! (.getRequestURI e)))]
-                          (-> (exec {:handler (report-uri-handler* {:report-callback callback})
-                                     :uri "/csp-report"})
-                              :status))))
+    (test/is (= :undefined
+                (let [a! (atom :undefined)
+                      handler (reify HttpHandler (handleRequest [_ e] (reset! a! (.getRequestURI e))))]
+                  (exec {:handler (report-uri-handler* {:report-handler handler})
+                         :uri "/not-csp-report"})
+                  @a!))))
 
-  (test/is (= #_:status 500
-                        (let [callback (fn [^HttpServerExchange _] (throw (Exception. "Oops")))]
-                          (-> (exec {:handler (report-uri-handler* {:report-callback callback})
-                                     :uri "/csp-report"})
-                              :status))))
+  (testing "request body handling"
+
+    (test/is (= "{\"csp-report\":{}}"
+                (let [a! (atom :undefined)
+                      handler (with-request-body-string (fn [s] (reset! a! s)))]
+                  (exec {:handler (report-uri-handler* {:report-handler handler})
+                         :uri "/csp-report"
+                         :method :post :body "{\"csp-report\":{}}"})
+                  @a!))))
+
+  (testing "response status"
+
+    (test/is (= #_:status 200
+                          (let [a! (atom :undefined)
+                                handler (reify HttpHandler (handleRequest [_ e] (reset! a! (.getRequestURI e))))]
+                            (-> (exec {:handler (report-uri-handler* {:report-handler handler})
+                                       :uri "/csp-report"})
+                                :status))))
+
+    (test/is (= #_:status 500
+                          (let [handler (reify HttpHandler (handleRequest [_ _] (throw (Exception. "Oops"))))]
+                            (-> (exec {:handler (report-uri-handler* {:report-handler handler})
+                                       :uri "/csp-report"})
+                                :status)))))
 
   )
 
 ;;,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
 
-(def ^{:arglists '([{:keys [policy, report-only, random-nonce-fn, report-callback]}])
+(def ^{:arglists '([{:keys [policy, report-only, random-nonce-fn, report-handler]}])
        :private true}
   csp-handler* (partial csp/csp-handler dummy-handler))
 
@@ -100,17 +125,30 @@
                                                  :random-nonce-fn (constantly "TEST-NONCE")})})
                   :headers (get "content-security-policy-report-only"))))
 
-  (testing ":report-callback"
+  (testing ":report-handler"
 
-    (test/is (= "report-uri /csp-report"
-                (-> (exec {:handler (csp-handler* {:policy {}
-                                                   :report-callback identity})})
-                    :headers (get "content-security-policy"))))
+    (testing "headers"
 
-    (test/is (= "report-uri /test-report-uri"
-                (-> (exec {:handler (csp-handler* {:policy {"report-uri" "/test-report-uri"}
-                                                   :report-callback identity})})
-                    :headers (get "content-security-policy"))))
+      (test/is (= "report-uri /csp-report"
+                  (-> (exec {:handler (csp-handler* {:policy {}
+                                                     :report-handler (handler/with-exchange identity)})})
+                      :headers (get "content-security-policy"))))
+
+      (test/is (= "report-uri /test-report-uri"
+                  (-> (exec {:handler (csp-handler* {:policy {"report-uri" "/test-report-uri"}
+                                                     :report-handler (handler/with-exchange identity)})})
+                      :headers (get "content-security-policy")))))
+
+    (testing "request body"
+
+      (test/is (= "{\"csp-report\":{}}"
+                  (let [a! (atom :undefined)
+                        report-handler (with-request-body-string (fn [s] (reset! a! s)))]
+                    (exec {:handler (csp-handler* {:policy {}
+                                                   :report-handler report-handler})
+                           :uri "/csp-report"
+                           :method :post :body "{\"csp-report\":{}}"})
+                    @a!))))
 
     )
 
@@ -119,7 +157,7 @@
 ;;,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
 
 (defn- get-request-nonce*
-  {:arglists '([{:keys [policy, report-only, random-nonce-fn, report-callback]}])}
+  {:arglists '([{:keys [policy, report-only, random-nonce-fn, report-handler]}])}
   [opts]
   (-> (reify HttpHandler
         (handleRequest [_ e]